another improvements
has included other soundsystem files
starts changing char pointer type into CString (more flexible)
removed many compiler warnings

+++ maszyna/include/sysinfo.h
@@ -113,7 +113,7 @@ inline void SetProcessorAffinity()
 
 inline double CALCULATEFPS()
 {
-	double newTime = glfwGetTime();
+	float newTime = float(glfwGetTime());
 
 	Global::frameTime = newTime - Global::previousFrameTime;
